requirements.txt
asgiref==3.5.2
astroid==2.12.13
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
dill==0.3.6
Django~=4.2.1
django-cors-headers==3.13.0
django-filter==22.1
django-phonenumber-field==7.0.1
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.4
isort==5.11.2
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.2
openapi-codec==1.3.2
phonenumbers==8.13.2
Pillow==9.3.0
platformdirs==2.6.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pylint==2.15.8
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
requests~=2.30.0
requests-oauthlib==1.3.1
simplejson==3.18.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomlkit==0.11.6
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
wrapt==1.14.1
